Soviet equipment update
-2k more AA1
-200k more gun 0
-1k more AT1

Some addition equipment added so players make even larger defense in depths than previous. This is because Horst barbarossa can go pretty deep into Russia before Russia dies. The quality and quantity of inf micro drops drastically in 1942-1943 (from my observations). With more infantry, it will be easier for Soviet players to micro tanks instead of infantry. The Soviet Union begins to run out of manpower in 1943-1944, forcing it to go All Adults Serve if there are too many losses, so it shouldn't be a problem.